articles/examples/comparison/requirements.txt
# AgentCore Framework Comparison Example
# https://docs.aws.amazon.com/bedrock-agentcore/

# AgentCore SDK
bedrock-agentcore>=1.0.0

# Strands Agents (AWS native framework)
strands-agents>=0.1.0

# LangGraph (graph-based agent workflows)
langgraph>=0.2.0
langchain-aws>=0.2.0
langchain-core>=0.3.0

# Utilities
boto3>=1.34.0
python-dotenv>=1.0.0
